Patched ParamChecker for multiple templatefqs
ParamChecker utilizes a function checkFileExists. This function only worked for single string variables, as it used os.path.isfile. Changed to loop over a list of arguments (default format for argparse output when nargs is specified), and added a check for str arguments (default format for argparse outputs when nargs is NOT specified). This allows checkFileExists to check if files exist for both arguments with single and multiple allowed arguments, without running into the problem of trying to loop over a string.