Changed unique read filtering method.
getVariantReads method used set() to try to make a unique list, however objects in memory are unique, even if they contain the same data. Changed this method to instead check the readID and pair number of reads, and add the reads to a list if their ID and pair number haven't already been added. This is important, because if reads are not filtered in this way, the next filter will remove any reads that occur more than 2 times.