VaSeBuilder function changes
Changes to main VaSeBuilder function. Added a function to define VaSeBuilder debug messages to make VaSeBuilder function code more readable. Added checks to see if acceptor context is empty, and allows VaSeBuilder to continue even if the acceptor is empty. Currently, this involves adding a dummy set of 2 reads to the acceptor reads, then uses this same set if there are again no reads in the combined context window in the acceptor. Variant context is set equal to the donor context. Added some debug messages using the new method. Fixed some typos. Restructured several if-then and try-except blocks by implementing guard-clause structure to increase readability, decrease indentation, and decrease the chance of catching an unexpected error.