Added CryptKeeper hashing/encryption utils.
CryptKeeper.sh is a proprietary script that reads in a varcon file and generates random salts and hashes for the contained sample names using Argon2, storing them in an external file. CryptKeeper.sh then calls CryptKeeper.py to write a new varcon file with the hashes in place of the sample names. It also generates an encrypted copy of the hash table file, the encrypted encryption (Fernet) key used to encrypt the hash table, and the key pair used to encrypt the encryption key.