Removed type check in varcon-out parameter.
Used to check if a file with that name already existed, to prevent overwriting, but this causes conflict if trying to run with a pre-existing varcon file with the default name. This will always cause the check to fail, so you would have to specify a varcon-out name even though no varcon will be produced. So instead, users will just have to not overwrite their own files.