Disabled used_bam checking when making headers.
I'm not sure that this actually does anything now, but it might be messing up the checks when using pre-made varcons. Reading in the pre-made varcons doesn't include any info about the bams that were used, so I think this ends up being empty and causes all samples to be filtered out when making the headers.